iterate through possible config file locations
incorporate other PR comments from Vic:
1. use os.path.join properly
2. put config.yml in /usr/share/calyptos

this also copies config.yml to the ~/.calyptos directory
if there IS NOT a calyptos RPM package installed. this
indicates that the user has used setup.py to install as
opposed to rpm/yum, and we are allowing for the possibility
that in a dev environment they may want to edit the config.