Reorder mido recipes in the calyptos phases
We will now try to run cassandra/zookeeper recipes without
clearing the runlist in between if the two components are
on the same host.

Move midonet-api to the bootstrap phase, which will also now
run mn-conf instead of being run in the create-first-resources
recipe.

This needs to come in parallel with midokura-cookbook changes (in
commit 193cb7).